requirements.txt
flask
flask_sqlalchemy
authlib
pyjwt
requests
gunicorn
python-dotenv
flasgger
flask-cors
flask-limiter
pytest


